Added background-position for 'default'
Problem state: The privacy policy page is not of category "technology,
community, careers" etc.  So the footer-doodle class is not accompanied
by a tag such as "technology" -- thus the background position of the
sprite is not known and not set.

New state: I have added a "background-position" (like a default) to the
sprite.  I have not tested this properly though I believe the theory is
a good one.